+++ src/main/java/de/omnikryptec/old/util/logger/SystemInputStream.java
@@ -30,7 +30,7 @@ import java.util.LinkedList;
 public class SystemInputStream {
 
     public static final char NEWLINECHAR = '\n';
-    public static final int NEWLINEINT = (int) NEWLINECHAR;
+    public static final int NEWLINEINT = NEWLINECHAR;
     public static final byte NEWLINEBYTE = (byte) NEWLINEINT;
 
     private final InputStream inputStreamOriginal;
